Create log key using sigsum-key
We take the opportunity to change the name of the keyfile to something
more appropriate -- now that the files generated are key and key.pub,
the name sk is even worse than before.

Note that the file format has changed, so simply renaming sk to logkey
for an already running log is not going to work.