Create the empty sth file on tree creation
If we've lost the sth key, a human is needed to sort out what tree to
create a new one from, so don't mask that error in
start_log(). Instead create the sth file when creating the tree.

Take the opportunity to change the name of the file where the tree id
is stored. It's not used by this role, nor by the log itself. Yet.