Fix how Go programs are installed
Makes it easier to install desired versions, both within the context
of a single user and across different users on the same system.  The
same default for "home" is used as before, but it is now tunable using
the sigsum_home variable.  Leave at its default value to avoid breakage.

This commit also ensures restarts are not triggered unless they are
needed.  "Needed" is defined as any of the relevant binaries changed.

Compatibility warning: some default variables on how to git-clone and
go-install have been deleted.  Tune install of each Go program using:

* sigsum_tools_from_sigsum_go
* sigsum_tools_from_log_go
* sigsum_tools_from_trillian_go

If you only specified Go versions of sigsum-go (sigsum_lib_version),
log-go (sigsum_log_version), and Trillian (sigsum_trillian_version), no
action is needed as the default versions are still set in the same way.

Note: the deletion of git as an apt-dependency in this commit means the
sigsum role is no longer bound to Debian (but we only test on Debian).