Revert sigsum_sth_file changes
An if statement was added in git-commit
856f15ad1f99f95cbe7e71756e602eefb423472c.  This has now been reverted,
because the variable is always given a constant in tasks/main.yml.

The variable was also added in defaults/main.yml in git-commit
a6c87a100f30c56d921d913275aa82d015532f2b.  This has now been reverted,
because any value that the user sets is overwritten in tasks/main.yml.