sigsum_agent: Start properly at boot
With WantedBy=multi-user.target things work, as opposed to when they
are set to sockets.target (for the socket unit) and
default.target (for the service unit).

sockets.target doesn't make sense since, according to systemd.socket(5):

    Socket units automatically gain a Before= dependency on sockets.target.

default.target seems wrong since it is an alias for
graphical.target (see output of systemctl status default.target).