Use tasks instead of handlers for litewitness
The purpose of this change is to make the litewitness role work
properly also when it is used to setup two or more litewitness
instances running in parallel on the same machine.

Handlers did not work well in that case because we do not have
enough control over when handlers are run and what values
different variables will have at that time. For example, it can
happen that when two witnesses foo and bar are to be configured,
the handler that was supposed to start the service for foo is
instead starting the service for bar, so that foo is missed.